Update sync endpoint to attack backlog from both ends
I decided to split this up into two queries because there may be a time/definitely was a time that the bot stopped posting to reddit for an extended period but was still processing update API requests. Once it comes back online, there is a massive queue of posts that need processing. During times of high volume of submissions, this can quickly overwhelm the bot and result in hours if not days before it gets caught up and starts mirroring new submissions. The compromise is to attack 5 of the oldest and 5 of the newest, ensuring the backlog is processed while new comments are, too. Eventually the bot will catch back up and things will work as expected (e.g. within 60 seconds)